Finished version of lab2.c
I originally had some issues with the numbers printing out random numbers and the number of hot, pleasant, and cold days, not showing up correctly at the end. I figured out that you have to initialize the variables with an actual number because the program doesn't automatically assume that an int with no specified number is 0.